moved email structs into subfolder
moved email_params and smtp_account dataclasses into subfolder email
