+++ lib/structs/email/email_params.py
