+++ lib/structs/email/smtp_account.py
